requirements.txt
asgiref==3.7.1
backports.zoneinfo==0.2.1
Django==4.2.1
django-ckeditor==6.5.1
django-cors-headers==4.0.0
django-js-asset==2.0.0
django-taggit==4.0.0
django-taggit-serializer==0.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.5.0
PyJWT==2.7.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.6.2
